deduplicate build code
