+++ scripts/build-image.py
@@ -2,16 +2,9 @@
 
 import os
 from argparse import ArgumentParser
-from shutil import copyfile, copytree, rmtree
+from shutil import copytree
 import docker
-import yaml
-
-BASE_DIRECTORY = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
-
-placeholder_values = {
-    "--COPY--": ["COPY lib /src/lib/"],
-    "--COMMAND--": ['CMD ["python", "server.py"]'],
-}
+import helper
 
 
 def parse_arguments():
@@ -28,34 +21,15 @@ def parse_arguments():
     return args
 
 
-def replace_lines(infile, outfile, replace_dict):
-    for line in infile:
-        key = line.rstrip()
-        if key in replace_dict:
-            outfile.write("\n".join(replace_dict[key]))
-        else:
-            outfile.write(line)
-
-
-def process_dockerfile(docker_env):
-    src_dockerfile = os.path.join(BASE_DIRECTORY, "docker", docker_env, "Dockerfile.template")
-    dest_dockerfile = "./Dockerfile"
-    print(f"\nCopying Dockerfile from {src_dockerfile} to {dest_dockerfile} ...")
-    with open(src_dockerfile, "r") as infile, open(dest_dockerfile, "w") as outfile:
-        replace_lines(infile, outfile, placeholder_values)
-    return dest_dockerfile
-
-
-def standard_copy(target_folder, docker_env):
-
-    dest_dockerfile = process_dockerfile(docker_env)
+def copy_files(target_folder, docker_env):
+    dest_dockerfile = helper.process_dockerfile(docker_env, helper.PLACEHOLDER_VALUES)
 
     src_lib = target_folder
     dest_lib = "./lib"
     print(f"\nCopying target dir from {src_lib} to {dest_lib} ...")
     copytree(src_lib, dest_lib)
 
-    copies = [dest_lib]
+    copies = [dest_lib, dest_dockerfile]
     return [os.path.abspath(file_copy) for file_copy in copies]
 
 
@@ -64,18 +38,13 @@ if __name__ == "__main__":
     cwd = os.getcwd()
 
     args = parse_arguments()
+    docker_env = args.docker_env
     config_file = args.config_file
     docker_env = args.docker_env
+    docker_image = args.image_name
 
     if config_file is None:
-        target_folder = os.path.abspath(args.target_folder)
-
-        if args.mdai_folder is None:  # If None, defaults to .mdai directory of target folder
-            mdai_folder = os.path.join(target_folder, ".mdai")
-        else:
-            mdai_folder = os.path.abspath(args.mdai_folder)
-
-        config_path = os.path.join(mdai_folder, "config.yaml")
+        target_folder, mdai_folder, config_path = helper.get_paths(args)
 
         # Detect config file if exists
         if os.path.exists(config_path):
@@ -83,42 +52,16 @@ if __name__ == "__main__":
 
     # Prioritize config file values if it exists
     if config_file is not None:
-        config_file = os.path.abspath(config_file)
-        parent_dir = os.path.dirname(config_file)
-        os.chdir(parent_dir)
+        docker_env, target_folder, mdai_folder = helper.process_config_file(config_file)
 
-        with open(config_file, "r") as stream:
-            data = yaml.safe_load(stream)
-            docker_env = data["base_image"]
-            target_folder = os.path.abspath(data["paths"]["model_folder"])
-            mdai_folder = os.path.abspath(data["paths"]["mdai_folder"])
-
-        os.chdir(cwd)
-
-    docker_image = args.image_name
     relative_mdai_folder = os.path.relpath(mdai_folder, target_folder)
-    os.chdir(os.path.join(BASE_DIRECTORY, "mdai"))
-
-    copies = standard_copy(target_folder, docker_env)
+    os.chdir(os.path.join(helper.BASE_DIRECTORY, "mdai"))
+    copies = copy_files(target_folder, docker_env)
 
     try:
-        print(f"\nBuilding docker image {docker_image} ...\n")
-        build_dict = {"MDAI_PATH": relative_mdai_folder}
-        response = client.api.build(
-            path=".", tag=docker_image, quiet=False, decode=True, buildargs=build_dict
-        )
-        for line in response:
-            if list(line.keys())[0] in ("stream", "error"):
-                value = list(line.values())[0]
-                if value:
-                    print(value.strip())
+        helper.build_image(client, docker_image, relative_mdai_folder)
     except docker.errors.APIError as e:
         print("\nBuild Error: {}".format(e))
     finally:
-        print("\nRemoving copied files...")
-        for file_copy in copies:
-            if os.path.isdir(file_copy):
-                rmtree(file_copy)
-            else:
-                os.remove(file_copy)
+        helper.remove_files(copies)
         os.chdir(cwd)
